requirements.txt
sarvamai
python-dotenv
sounddevice
soundfile
numpy
